Added Photon Weighting
Photon weighting allows us to simulate probabilistic removal (like a detector's quantum efficiency or a grating's reflectivity) with fewer photons. Each photon is given a weight that signifies the probability it made it through the instrument.

Also:
Wolter Optic Objects now automatically remove photons which miss, this behavior is consistent with that of other types of objects.