Improved Astropy Unit Checking
Added decorators to functions that check if the arguments are astropy units that can be converted to the proper type. Previously we had basic conditionals checking that the arguments were u.quantity.Quantity's. We were unable to check if the input units could be converted to the desired units, but now we can!