Fixed Rays Addition
When two Rays objects are added such that the first has wavelength and order while the other does not, now the second will be given wavelength and order equal to the mean wavelength and order of the first. This is not an ideal solution but it is strongly recommended that the user only add Rays when they are sure that they both have waves/orders or neither of them do.